API/picknpay/requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.7
django-filter==2.4.0
django-framework==0.1.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
Pillow==8.1.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
Werkzeug==1.0.1
